mpid-server/pyproject.toml
[project]
name = "mpid-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "uvicorn>=0.35.0",
]
